rename to hc and cleanup code
