--- holoconsole/header.txt
@@ -1,51 +1,0 @@
-                                           &
-                                 @   #%%     ( #
-                   % #     #  #(&#(&### (#(  #& @#(       /(
-                       # ((&# @# #(##&&# /#%## &    %% %
-              #      %# ####%##( ( (#(((((((%&((&(&(      # #(   &
-                 # # # &#%%@## ((((((((((((((((((((((%%(  (##
-                &  % #  #%##%################(((((((((((# (( (#(
-               (  #  & ######%%%#              (##((((((((     @
-             % # #*# #(((##%%                      ##(((((( (((      &
-        (   /  (&#%((#((##%                          ##((((#*# #      (
-            # (  (#(((((#                             .#(((# ###(
-         ((/&   (%///((#                                #(((((# # ##%
-           ( (/#//#//(#                                  #((((&##@ (
-          * /(/(( ///((                                  &##(#(@
-           ////#( //((                                    %##%( &&#
-           ** / /////(                                    %### # ##
-        / (      ////(                                    %%### ##    #
-             @%//////(                                   ,%#####  @
-         /(   /   ////(                                  %####  % # &
-        &/*//#/ ///*//(#                                %##(((  #    %#
-              / /#/////(#                              %#((((& (&
-            (#// //*/////(                            ##((((%%   #
-            / * / *%%#////((                        ##((//&#  (   #%
-                     //*/////(#                  %#((//#//%( (
-                  /*  /@////////((#/        (##((/////#%( & #
-             (         &/////////////////////////////(/% %      * #
-            /     /&#   // // /////////////****//*// /&(( /%      %
-                   /  # &/    &*#/ /////  // /*///  / / *   /
-                          *# ( (&& ( /*/ / #/#//(&/   /(
-                            * & /* @/  /#   /*    / #
-                                */   / / /    /        /
-                                  /     /
-
-================================================================================
-======================   Holochain Console Container   =========================
-================================================================================
-
-Welcome to the Holochain CLI Container.
-This is a full container implementation that allows youy to install and
-instantiate multiple Holochain Genomes and provides means to call zome functions
-in active Holochain instances.
-
-To get started, try:
-
-var genome_hashes = Container.installedApps()
-var genome_instance = Container.instantiate(genome_hashes[0])
-console.log("Found genome: " + genome_instance.name())
-genome_instance.start()
-genome_instance.call("zome_name", "capability_name", "function_name", "{\"parameter\":\"test\"}")
-...
-
