--- holoconsole/qml.qrc
@@ -1,6 +1,0 @@
-<RCC>
-    <qresource prefix="/">
-        <file>shims.js</file>
-        <file>header.txt</file>
-    </qresource>
-</RCC>
